allow decimal values to be entered for rate (but cast it back to an integer anyway)

+++ TTSButSpecificAudioDevice/AudioPlaybackEngine.cs
@@ -110,7 +110,7 @@ internal class AudioPlaybackEngine : IDisposable
         using MemoryStream memoryStream = new();
         
         Synthesizer.SelectVoice(Voices[speechMessage.Voice]);
-        Synthesizer.Rate = speechMessage.Rate;
+        Synthesizer.Rate = (int)speechMessage.Rate;
         Synthesizer.SetOutputToAudioStream(memoryStream, SpeechFormat);
         Synthesizer.Speak(speechMessage.Text);
